Improving `draw_sample`  (#604)
* * optionally add a custom `.ttf` font file for plots
* when using custom font, draw rect using PIL, not cv2
* optional `color_map` to use specific color for labels
* utility function to convert color representation
* expose mask `blend` parameter
* `display_score` option
* `prettify` for nicer label rendering

* silly typing bugfix

* fix faulty default option

* truncate to 2 decimal points when pretty printing

* expose `erode_strength` for finer mask viz control; option to return as array / PIL Image

* imports

* mask border thickness

* conform to requested return type

* `denormalise_mask` for plotting nothing but masks

* better defaults -- prettify & show scores

* add scope for custom `prettify` function

* export `as_rgb_tuple` via `__all__`

* move `draw_bbox` from cv2 -> PIL

* correct if statement

* sketch `draw_keypoints`

* sketch `draw_mask` with PIL

* * add params for `draw_sample` wherever it's called
* font_scale -> font_size
* font -> font_path

* * add missing `display_label`
* remove last bit of `cv2` code

* patch `show_*` functions with kwargs (ugh)

* download & load default font

* * get_id -> get_by_id
* patch `label_str==""` when no class_map is given

* black formatting

* fix conflict

* detect -> detection

* git is complicated...

* fix rebase issues...

* new API compatibility

* revise font URL

* ensure dummy array is of type `np.uint8`

* fix accidental rebase error

* update changelog

+++ tests/visualize/test_draw_data.py
@@ -28,7 +28,7 @@ def test_draw_pred():
 
 
 def test_draw_keypoints(keypoints_img_128372):
-    img = np.zeros((427, 640, 3))
+    img = np.zeros((427, 640, 3)).astype(np.uint8)
     color = (np.random.random(3) * 0.6 + 0.4) * 255
     kps = KeyPoints.from_xyv(keypoints_img_128372, COCOKeypointsMetadata)
     img = draw_keypoints(img=img, kps=kps, color=color)
